implemented the mobile menu
